[chore] #192 Feature 단위 이벤트 분리
