+++ core/analytics/src/main/kotlin/com/neki/android/core/analytics/event/AnalyticsEvent.kt
@@ -1,0 +1,8 @@
+package com.neki.android.core.analytics.event
+
+sealed interface AnalyticsEvent {
+
+    val name: String
+    val params: Map<String, String>
+        get() = emptyMap()
+}
