+++ core/analytics/src/main/kotlin/com/neki/android/core/analytics/event/GlobalAnalyticsEvent.kt
@@ -1,0 +1,8 @@
+package com.neki.android.core.analytics.event
+
+sealed interface GlobalAnalyticsEvent : AnalyticsEvent {
+
+    data object AppOpen : GlobalAnalyticsEvent {
+        override val name = "app_open"
+    }
+}
